manager/backend/pyproject.toml
[project]
name = "chatgpt-manager-backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "httpx>=0.28",
    "python-multipart>=0.0.26",
    "curl-cffi>=0.15.0",
    "pyotp>=2.9",
    "qrcode[pil]>=8.0",
    "tortoise-orm>=1.1.7",
    "aiosqlite>=0.22.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
